+++ packages/bosia/templates/drizzle/src/features/todo/index.ts
@@ -1,3 +1,0 @@
-export * from "./schemas/todo.table";
-export * from "./types";
-export * as todoQueries from "./queries";