+++ packages/bosia/templates/drizzle/src/lib/components/todo/index.ts
@@ -1,3 +1,0 @@
-export { default as TodoForm } from "./todo-form.svelte";
-export { default as TodoItem } from "./todo-item.svelte";
-export { default as TodoList } from "./todo-list.svelte";