+++ registry/features/todo/todo-index.ts
@@ -1,3 +1,4 @@
 export * from "./schemas/todo.table";
 export * from "./types";
-export * as todoQueries from "./queries";
+export { TodoRepository } from "./todo.repository";
+export { TodoService } from "./todo.service";